pyproject.toml
[project]
name = "choptimize"
version = "0.1.0"
description = "CLI tool for analysing prompt quality in developer-LLM code conversations"
readme = "README.md"
authors = [{ name = "darragh0" }]
license = "MIT"
keywords = ["prompt-quality", "code-generation", "llm", "analysis"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
requires-python = ">=3.13"
dependencies = [
  "common",
  "annotated-types>=0.7.0",
  "datasets>=4.5.0",
  "fasttext-predict>=0.9.2.4",
  "matplotlib>=3.10.0",
  "numpy>=2.0.0",
  "openai>=1.0.0",
  "pandas>=2.0.0",
  "radon>=6.0.0",
  "rich>=14.3.1",
  "rich-argparse>=1.6.0",
  "scipy>=1.15.0",
  "seaborn>=0.13.0",
  "fastapi>=0.135.2",
  "uvicorn>=0.34.0",
  "python-multipart>=0.0.20",
  "jinja2>=3.1.6",
  "chromadb>=1.0.0",
  "sentence-transformers>=4.1.0",
  "google-genai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/darragh0/choptimize"
Repository = "https://github.com/darragh0/choptimize?tab=readme-ov-file#choptimize"

[dependency-groups]
dev = [
  "basedpyright>=1.38.4",
  "deptry>=0.24.0",
  "mypy>=1.19.1",
  "pandas-stubs>=3.0.0.260204",
  "pyright>=1.1.408",
  "ruff>=0.14.14",
  "scipy-stubs>=1.17.1.3",
  "typ>=0.11.0",
]

[tool.basedpyright]
typeCheckingMode = "standard"
reportUnknownMemberType = false           # incomplete matplotlib/seaborn stubs
reportUnusedCallResult = false            # unused matplotlib setters return Text objects
reportExplicitAny = false                 # Required for dict[str, Any]
reportInvalidStringEscapeSequence = false # Allow raw Unicode chars in display strings

[tool.deptry]
known_first_party = ["common", "utils", "preproc", "download", "analysis"]

[tool.uv.sources]
common = { workspace = true }

[tool.uv.workspace]
members = ["common"]
